refactor index into separate modules
