+++ .gitignore
@@ -21,3 +21,4 @@ MovieClub.xcodeproj/xcuserdata/marcus.xcuserdatad/xcdebugger/Breakpoints_v2.xcbk
 MovieClub.xcodeproj/xcuserdata/marcus.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
 *.xcodeproj
 ./MovieClub.xcodeproj
+node_modules/
